refactor: test init
